# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Presences API",
        "description": "Cara login:\r\n * Login ke endpoint /auth/login dgn email dan password Anda\r\n * Kopi token dari resultnya\r\n * Klik button authorize di pojok kanan atas\r\n * Masukkan token, dengan format::  Bearer <api_key>\r\n * Klik submit button Authorize",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/",
            "description": "LIVE server"
        }
    ],
    "paths": {
...
        "/workunits": {
            "get": {
                "tags": [
                    "Workunits"
                ],
                "summary": "Retrieve workunit list",
                "description": "Retrieve workunit list",
                "operationId": "d9f82c5465d33bfa5ff937143aa64fa2",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "example": "10"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "example": "1"
                        }
                    },
                    {
                        "name": "sort_by",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "id",
                                "created_at"
                            ]
                        }
                    },
                    {
                        "name": "order_by",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "workunit list"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/workunits/{id}": {
            "get": {
                "tags": [
                    "Workunits"
                ],
                "summary": "Retrieve workunit list",
                "description": "Retrieve workunits detail",
                "operationId": "ffe9991397d1f8b00f63d760236ea871",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "workunit detail"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/workunits/{id}/worktime": {
            "delete": {
                "tags": [
                    "Workunits"
                ],
                "summary": "delete Worktime",
                "description": "Workunit delete worktime",
                "operationId": "de9039491c281d662b3eece4186ae7ef",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/WorktimeId"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "delete worktime"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Workunits"
                ],
                "summary": "Assign worktimes on workunits",
                "description": "Assign worktimes on workunits",
                "operationId": "a1ac06a5f6eee891ce154c4a2ba22579",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/WorktimeId"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Workunit updated"
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "WorktimeId": {
                "properties": {
                    "worktime_id": {
                        "type": "integer",
                        "example": "1"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "Bearer": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}